decoder: set target encoding to UTF-8, not native
With the previously set option of CE_NATIVE, "latin1" would not be a
safe default on a system with a _different_ single-byte encoding (e.g.
CP-1251). The decoding would fail, and the workaround would be to set
the file encoding to the native encoding of the current R session, which
results in mojibake.

Instead, we can decode into UTF-8 (which should "just work" in almost
all cases) and let R handle any problems related to the session
encoding. This lets us decode latin1 on a CP-1251 system and the other
way around.