tree=TRUE: don't crash on unsupported objects
Previously, if an .opj file contained unsupported objects, importing it
with tree = TRUE would result in an error in the R code resulting from a
failed lookup by unique name.

NULLs in the tree structure, which correspond to such unsupported
objects, are now skipped. The name lookup function has been hardened
against similar problems in the future. The tree.opj example now
contains an unsupported graph object in order to test this.