Work around a compilation error in tree.hh
tree.hh depends on ptrdiff_t being available in the global namespace
without #including the right files that guarantee that to happen. On
modern compilers, <cstddef> (included by one of tree.hh's dependencies)
provides the type definition, but it's missing on GCC 6.3.

Restructure the build so that <stddef.h> is always #included before
tree.hh. (Something in the R headers included by Rcpp seems to include
<stddef.h>, so no changes in read_opj.cpp.)