read.opj: force native encoding of the file path
When accepting an std::string, Rcpp takes the CHAR() pointer in order to
populate the contents of the object. This will cause problems when
giving a string properly marked as UTF-8 to the code in an R session
where the native encoding is not UTF-8 (e.g. old Windows).

Force the conversion on the R side and document the requirement. No
regression tests because it's hard to guarantee any non-ASCII path to
work on machines running R CMD check.

Document this change and the previous one in the NEWS.